+++ benchmark.py
@@ -8,14 +8,16 @@ from functools import partial
 from pymongo import MongoClient
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, scoped_session
+from redis import Redis
 
 from vakt import (
     MemoryStorage, DENY_ACCESS, ALLOW_ACCESS,
-    Policy, RegexChecker, RulesChecker, Guard, Inquiry,
+    Policy, RegexChecker, StringExactChecker, StringFuzzyChecker, RulesChecker, Guard, Inquiry,
 )
 from vakt.storage.mongo import MongoStorage
 from vakt.storage.sql import SQLStorage
 from vakt.storage.sql.migrations import SQLMigrationSet
+from vakt.storage.redis import RedisStorage, JSONSerializer, PickleSerializer
 from vakt.rules import operator, logic, list, net
 
 
@@ -29,7 +31,7 @@ similar_regexp_policies_created = 0
 parser = argparse.ArgumentParser(description='Run vakt benchmark.')
 parser.add_argument('-n', '--number', dest='policies_number', nargs='?', type=int, default=100000,
                     help='number of policies to create in DB (default: %(default)d)')
-parser.add_argument('-s', '--storage', choices=('mongo', 'memory', 'sql'), default='memory',
+parser.add_argument('-s', '--storage', choices=('mongo', 'memory', 'sql', 'redis'), default='memory',
                     help='type of storage (default: %(default)s)')
 parser.add_argument('-d', '--dsn', dest='sql_dsn', nargs='?', type=str, default='sqlite:///:memory:',
                     help='DSN connection string for sql storage (default: %(default)s)')
@@ -43,6 +45,10 @@ regex_group.add_argument('--same', type=int, default=0,
                          help='number of similar regexps in Policy')
 regex_group.add_argument('--cache', type=int,
                          help="number of LRU-cache for RegexChecker (default: RegexChecker's default cache-size)")
+
+redis_group = parser.add_argument_group('Redis Storage related')
+redis_group.add_argument('--serializer', choices=('json', 'pickle'), default='json',
+                         help='type of serializer for policies stored in Redis (default: %(default)s)')
 
 ARGS = parser.parse_args()
 
@@ -124,6 +130,10 @@ def gen_policy():
 def get_checker():
     if ARGS.checker == 'rules':
         return RulesChecker()
+    elif ARGS.checker == 'exact':
+        return StringExactChecker()
+    elif ARGS.checker == 'fuzzy':
+        return StringFuzzyChecker()
     return RegexChecker(ARGS.cache) if ARGS.cache else RegexChecker()
 
 
@@ -181,6 +191,18 @@ def get_storage():
         # todo - why is there left an uncommitted transaction?
         sql_session.commit()
         migration.down()
+    if ARGS.storage == 'redis':
+        collection = 'vakt_policies_benchmark'
+        client = Redis('127.0.0.1', 6379, db=0)
+        if ARGS.serializer == 'json':
+            serializer = JSONSerializer()
+        elif ARGS.serializer == 'pickle':
+            serializer = PickleSerializer()
+        else:
+            serializer = None
+        yield RedisStorage(client, collection=collection, serializer=serializer)
+        client.flushdb()
+        client.close()
     else:
         yield MemoryStorage()
 
@@ -199,6 +221,7 @@ if __name__ == '__main__':
         print('Number of unique Policies in DB: {:,}'.format(overall_policies_created))
         print('Among them Policies with the same regexp pattern: {:,}'.format(similar_regexp_policies_created))
         print('Checker used: %s' % checker.__class__.__name__)
+        print('Storage used: %s' % st.__class__.__name__)
         # print('Inquiry looks like: %s' % vars(inq))
         print('Decision for 1 Inquiry took: %0.4f seconds' % (stop - start))
         print('Inquiry passed the guard? %s' % allowed)